(Pulled from #1461, do not merge) Factor out shared methods into common module
