+++ components/places/src/import/mod.rs
@@ -2,5 +2,6 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
+pub mod common;
 pub mod ios_bookmarks;
 pub use ios_bookmarks::import_ios_bookmarks;
